Refactor Slotmachine
